Relocate TraceBox and remove rc.rs
