+++ differential-dataflow/src/trace/wrappers/mod.rs
@@ -3,4 +3,3 @@
 pub mod enter;
 pub mod enter_at;
 pub mod frontier;
-pub mod rc;
